+++ app/src/main/java/com/codesoom/assignment/controllers/TaskController.java
@@ -1,8 +1,10 @@
 package com.codesoom.assignment.controllers;
 
+import com.codesoom.assignment.application.TaskRepository;
 import com.codesoom.assignment.exception.ParameterEmptyException;
 import com.codesoom.assignment.exception.TaskIdNotFoundException;
 import com.codesoom.assignment.model.Task;
+import org.springframework.context.annotation.DependsOn;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -28,22 +30,28 @@ import java.util.List;
 @RequestMapping("/tasks")
 @CrossOrigin
 public class TaskController {
-    private List<Task> tasks = new ArrayList<>();
+
+    private TaskRepository taskRepository;
+
+   public TaskController() {
+       taskRepository = new TaskRepository();
+    }
 
     @GetMapping
     public List<Task> taskList() {
-        return tasks;
+        return taskRepository.getTasks();
     }
 
     @GetMapping("/{id}")
     public Task taskDetail(@PathVariable Long id) {
-        return findTask(id);
+        System.out.println(this.taskRepository);
+        return taskRepository.getTask(id);
     }
 
     @PostMapping
     public ResponseEntity<Task> taskCreate(@RequestBody Task task) throws Throwable {
-        task.setId(generatedId());
-        tasks.add(task);
+
+        taskRepository.createTask(task);
         URI uri = ServletUriComponentsBuilder
                 .fromCurrentRequest()
                 .path("/{id}")
@@ -60,44 +68,14 @@ public class TaskController {
 
     @RequestMapping(path = "/{id}", method = {RequestMethod.PATCH, RequestMethod.PUT})
     public Task taskModify(@PathVariable Long id, @RequestBody Task task) {
-        validateParameter(id, task);
-        Task foundTask = findTask(id);
-        int foundTaskIndex = tasks.indexOf(foundTask);
-        task.setId(foundTask.getId());
-        tasks.set(foundTaskIndex, task);
-        return task;
+        return taskRepository.updateTask(id,task);
+
     }
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void taskDelete(@PathVariable Long id) {
-        Task findTask = findTask(id);
-        tasks.remove(findTask);
+        taskRepository.removeTask(id);
     }
 
-    private Task findTask(Long id) {
-        return tasks.stream()
-                .filter(task -> task.getId().equals(id))
-                .findFirst()
-                .orElseThrow(TaskIdNotFoundException::new);
-    }
-
-    private Long generatedId() throws Throwable {
-        Long maxId = 1L;
-        if (tasks.size() > 0) {
-            Comparator<Task> comparator = Comparator.comparingLong(Task::getId);
-            maxId = tasks.stream()
-                    .max(comparator)
-                    .orElseThrow(TaskIdNotFoundException::new)
-                    .getId() + 1;
-        }
-        return maxId;
-    }
-
-    private void validateParameter(Long id, Task task) {
-        final String newTitle = task.getTitle();
-        if (id == null || newTitle.isBlank()) {
-            throw new ParameterEmptyException(id, newTitle);
-        }
-    }
 }
